+++ mangadex_library/test/mangadex_library_example.dart
@@ -8,8 +8,8 @@ void main() {
 void printFilenames() async {
   // this function, needs a mangadex account username and password supplied
   // to retrive login token
-  var username = 'USERNAME'; // Put your username here
-  var password = 'PASSWORD'; // Put your password here
+  var username = 'riksantra'; // Put your username here
+  var password = 'Sikkim123.'; // Put your password here
 
   //The line below uses the login function and takes in
   //two String parameters, username and password and returns
@@ -36,23 +36,25 @@ void printFilenames() async {
 
   var chapterID = chapterData!
       .data[0].id; // This line sets the chapterID variable to the chapter id of
-  // the first chapter from the chapterData we just got.
-  //Every chapter has a usique chapter ID and a chapter Hash
-  //Chapter ID is required to access info of the desired chapter.
-  var chapterHash = chapterData.data[0].attributes
-      .hash; // this variable stores the chapter hash of a chapter
-  //Chapter Hash is required for requesting manga pages.
+  // // the first chapter from the chapterData we just got.
+  // //Every chapter has a usique chapter ID and a chapter Hash
+  // //Chapter ID is required to access info of the desired chapter.
+  // var chapterHash = chapterData.data[0].attributes
+  //     .hash; // this variable stores the chapter hash of a chapter
+  // //Chapter Hash is required for requesting manga pages.
 
-  var baseUrl = await lib.getBaseUrl(
-      chapterID); // This variable stores the baseUrl(Authority) to the chapter we are looking for
-  //BaseUrl always require a chapter ID to obtain an address.
-  var filenames = await JsonUtils.getChapterFilenames(chapterID, false);
+  // var baseUrl = await lib.getBaseUrl(
+  //     chapterID); // This variable stores the baseUrl(Authority) to the chapter we are looking for
+  // //BaseUrl always require a chapter ID to obtain an address.
+  // var filenames = await JsonUtils.getChapterFilenames(chapterID, false);
 
-  // the filenames variable stores the name of all files in a manga chapter
-  // using the getChapterFileNames function provided in the jsonSearchCommands.dart file.
-  for (var i = 0; i < filenames.length; i++) {
-    print(lib.constructPageUrl(
-        chapterID, token, chapterHash, filenames[i], false));
-  }
-  // this for loop prints the url to all the pages in the provided chapters.
+  // // the filenames variable stores the name of all files in a manga chapter
+  // // using the getChapterFileNames function provided in the jsonSearchCommands.dart file.
+  // for (var i = 0; i < filenames.length; i++) {
+  //   print(lib.constructPageUrl(
+  //       chapterID, token, chapterHash, filenames[i], false));
+  // }
+  // // this for loop prints the url to all the pages in the provided chapters.
+  var result = await lib.markChapterRead(token, chapterID);
+  print(result.result);
 }